Convert configure to cpp
In the old implementation the filepath of the config was send. This was needed as in python we can't resolve all our custom tue_config stuff. But this can cause issues, when ed is running on another machine than the configure node. By converting the configure node to cpp, the file is parsed locallaly include our custom stuff. As we can send the a yaml string again.